+++ kitty/main.py
@@ -15,7 +15,7 @@ from .child import set_default_env
 from .cli import create_opts, parse_args
 from .cli_stub import CLIOptions
 from .conf.utils import BadLine
-from .config import cached_values_for, initial_window_size_func
+from .config import cached_values_for, initial_window_size_func, expandvars
 from .constants import (
     appname, beam_cursor_data_file, config_dir, glfw_path, is_macos,
     is_wayland, kitty_exe, logo_data_file, running_in_kitty
@@ -230,6 +230,22 @@ def get_editor_from_env(shell_env: Mapping[str, str]) -> Optional[str]:
                 return editor
 
 
+def expand_listen_on(listen_on: str, from_config_file: bool) -> str:
+    listen_on = expandvars(listen_on)
+    if '{kitty_pid}' not in listen_on and from_config_file:
+        listen_on += '-{kitty_pid}'
+    listen_on = listen_on.replace('{kitty_pid}', str(os.getpid()))
+    if listen_on.startswith('unix:'):
+        path = listen_on[len('unix:'):]
+        if not path.startswith('@'):
+            if path.startswith('~'):
+                listen_on = f'unix:{os.path.expanduser(path)}'
+            elif not os.path.isabs(path):
+                import tempfile
+                listen_on = f'unix:{os.path.join(tempfile.gettempdir(), path)}'
+    return listen_on
+
+
 def setup_environment(opts: OptionsStub, cli_opts: CLIOptions) -> None:
     extra_env = opts.env.copy()
     if opts.editor == '.':
@@ -241,7 +257,12 @@ def setup_environment(opts: OptionsStub, cli_opts: CLIOptions) -> None:
             os.environ['EDITOR'] = editor
     else:
         os.environ['EDITOR'] = opts.editor
-    if cli_opts.listen_on:
+    from_config_file = False
+    if not cli_opts.listen_on and opts.listen_on.startswith('unix:'):
+        cli_opts.listen_on = opts.listen_on
+        from_config_file = True
+    if cli_opts.listen_on and opts.allow_remote_control != 'n':
+        cli_opts.listen_on = expand_listen_on(cli_opts.listen_on, from_config_file)
         os.environ['KITTY_LISTEN_ON'] = cli_opts.listen_on
     set_default_env(extra_env)